Working on defining the problem leading to Zlib::DataError
Files of Interest

net folder contains the MRI-2.0.0 net-libs. They are prepended to path
of test files.

wunderground.rb is the main test file
zlib_targeted.rb is an attempt to isolate the error away from Net/HTTP
and focus on the Zlib error, possibly due to a corrupted gzip response
from webserver.

socket_content.bin is a binary dump of the gzipped HTTP response. It was
dumped by dropping a `binding.pry` inside of Net/HTTP/response.rb ~ line
360 (?).  Then the socket object was called `socket.read_all` and that
data was dumped to socket_content.bin.

Further exploration was done via curl to see what response the Webserver
gives. Of interest are the files curl_response and curl_raw.  Note that
line 12 has an artifact as follows:

f03

This artifact is not present in the standard (non-raw) dump from curl.

@zph

`tree .`
    .
    ├── 1_9_3_result.md
    ├── 2_0_0_result.md
    ├── README.md
    ├── curl_raw.txt
    ├── curl_response.txt
    ├── http-1.9.3.rb
    ├── net
    │   ├── ftp.rb
    │   ├── http
    │   │   ├── backward.rb
    │   │   ├── exceptions.rb
    │   │   ├── generic_request.rb
    │   │   ├── header.rb
    │   │   ├── proxy_delta.rb
    │   │   ├── request.rb
    │   │   ├── requests.rb
    │   │   ├── response.rb
    │   │   └── responses.rb
    │   ├── http.rb
    │   ├── https.rb
    │   ├── imap.rb
    │   ├── pop.rb
    │   ├── protocol.rb
    │   ├── smtp.rb
    │   └── telnet.rb
    ├── net-1.9.3-p429
    │   ├── ftp.rb
    │   ├── http.rb
    │   ├── https.rb
    │   ├── imap.rb
    │   ├── pop.rb
    │   ├── protocol.rb
    │   ├── smtp.rb
    │   └── telnet.rb
    ├── socket_content.bin
    ├── wunderground-working-2_0.rb
    ├── wunderground.rb
    └── zlib_targeted.rb

    3 directories, 35 files